requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.49.0
joblib==1.3.2
keyboard==0.13.5
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
pyttsx3==2.90
scikit-learn==1.2.2
scipy==1.12.0
six==1.16.0
threadpoolctl==3.3.0
